プログラムサイズの縮小を試行（Arduinoに書き込み不可）

+++ soil_moisture_sensor.h
@@ -12,9 +12,9 @@ const byte ADDRESS = 0x28;
 class SoilMoistureSensor {
    private:
     // センサーからの応答を取得する試行回数。
-    int _times;
+    byte _times;
     // センサーから応答の取得する間隔ミリ秒
-    int _interval;
+    byte _interval;
 
     // デフォルトコンストラクタ
     SoilMoistureSensor(void) {}
@@ -24,7 +24,7 @@ class SoilMoistureSensor {
     // Arguments:
     //  times: センサーからの応答を取得する試行回数。
     //  interval: センサーから応答の取得する間隔ミリ秒
-    SoilMoistureSensor(int times, int interval) {
+    SoilMoistureSensor(byte times, byte interval) {
         this->_times = times;
         this->_interval = interval;
     }
@@ -34,7 +34,7 @@ class SoilMoistureSensor {
     // Returns:
     //  センサーが接続されていることを確認できた場合はtrue。確認できなかった場合はfalse。
     bool init(void) {
-        for (int i = 0; i < this->_times; ++i) {
+        for (byte i = 0; i < this->_times; ++i) {
             Wire.beginTransmission(ADDRESS);
             if (Wire.endTransmission() == 0) {
                 return true;
@@ -52,7 +52,7 @@ class SoilMoistureSensor {
     //
     // Returns:
     //  土壌水分量。マイナスの値が返却された場合は測定エラー。
-    int32_t measure(void) {
+    int16_t measure(void) {
         // 測定する命令を発効
         Wire.beginTransmission(ADDRESS);
         Wire.write(COMMAND_GET_VALUE);
@@ -62,7 +62,7 @@ class SoilMoistureSensor {
         Wire.requestFrom(ADDRESS, (uint8_t)0x02);
 
         // 送信された測定値を読み込み
-        for (int i = 0; i < this->_times; ++i) {
+        for (byte i = 0; i < this->_times; ++i) {
             while (Wire.available()) {
                 uint8_t adc_value_l = Wire.read();
                 uint8_t adc_value_h = Wire.read();
@@ -73,10 +73,10 @@ class SoilMoistureSensor {
                     adc_value = 1023;
                 }
                 adc_value = 1023 - adc_value;
-                return (int32_t)((double)adc_value / 1023 * 100);
+                return (int16_t)((float)adc_value / 1023 * 100);
             }
             delay(this->_interval);
         }
-        return ~(int32_t)0;
+        return ~(int16_t)0;
     }
 };
